requirements.txt
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
graphviz==0.13.2
hdbscan==0.8.26
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
pandas==1.0.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
shap==0.35.0
six==1.14.0
sklearn==0.0
tqdm==4.45.0
xgboost==1.0.2
